refactor(form/render): Extrai lógica de renderização do formulário para hook
Move a lógica de estado, callbacks e processamento da página de visualização do formulário para o hook useFormRenderer. Isso melhora a modularidade e reusabilidade do código.
